requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==24.2.0
chardet==4.0.0
discord.py==1.7.3
environs==11.1.0
greenlet==3.1.1
idna==3.10
loguru==0.7.2
marshmallow==3.23.1
multidict==6.1.0
packaging==24.2
playwright==1.48.0
propcache==0.2.0
pyee==12.0.0
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.17.1
